refactor: Merge Gesture and TouchGesture
